requirements.txt
<<<<<<< HEAD
﻿attrs==25.3.0
beautifulsoup4==4.13.4
bing==0.31
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flask_sqlalchemy
gunicorn
googlesearch-python==1.3.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
pycparser==2.22
pyfiglet==1.0.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
psycopg2-binary
requests==2.32.4
requests-file==2.1.0
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tabulate==0.9.0
termcolor==3.1.0
tldextract==5.3.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
whois==1.20240129.2
wsproto==1.2.0
=======
﻿attrs==25.3.0
beautifulsoup4==4.13.4
bing==0.31
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flask_sqlalchemy
gunicorn
googlesearch-python==1.3.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
pycparser==2.22
pyfiglet==1.0.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
psycopg2-binary
requests==2.32.4
requests-file==2.1.0
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tabulate==0.9.0
termcolor==3.1.0
tldextract==5.3.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
whois==1.20240129.2
wsproto==1.2.0
>>>>>>> 4aac04f96e61deeb237f8c8cde28c29448d55b57
